Adding more debugging and fixing timeout for semaphore.
Passing on changes to new repo from previous repo. Making sure to add in a
sleep in the busy while() loop while waiting to hover so  we actually catch
the state change.

Added changes to setFlyingState so we can debug state changes. Added two states that Logan and I saw during
tuesday's testing. Add more states here so we know wtf is going on.

Added changes to moveRelativeMeters to add a passback bool for timeout of our command
since we're now using waitTimed() with a timeout of 10 seconds. This timeout can be
adjusted later with respect to distance and heading changes. TBD based on measurements taken.